Make repetition safe

+++ src/index.js
@@ -5,34 +5,43 @@
  */
 
 import * as p from 'path';
-import { writeFileSync } from 'fs';
+import { writeFileSync, existsSync, readFileSync } from 'fs';
 import { sync as mkdirpSync } from 'mkdirp';
+
+import {
+  createDescriptorsWithKey,
+  defineMessageFormat,
+  removeDefindeMessageFormat,
+} from './utils';
 
 const EXTRACTED_TAG = Symbol('DescriptorGenerated');
 
-export default function ({types: t}) { // eslint-disable-line no-unused-vars
+export default function({ types: t }) { // eslint-disable-line no-unused-vars
 
   function isTextIgnorable(text) {
     const hasWords = /^\w+/.test(text); // covers empty string
     const isNumber = /^\d+$/.test(text);
-    return (text.length === 1 || isNumber || !hasWords );
+    return text.length === 1 || isNumber || !hasWords;
   }
 
   function generateDescriptorFromText(defaultMessage, path, state) {
-    const { file, reactIntlMessages : { generatedDescriptors, proxyTexts } } = state;
+    const {
+      file,
+      reactIntlMessages: { generatedDescriptors, proxyTexts },
+    } = state;
 
     defaultMessage = defaultMessage.trim();
 
-    if (isTextIgnorable(defaultMessage) || proxyTexts.has(defaultMessage)) return;
+    if (isTextIgnorable(defaultMessage) || proxyTexts.has(defaultMessage))
+      return;
 
     proxyTexts.add(defaultMessage);
 
-    const prefixId = p.relative(
-      process.cwd(),
-      file.opts.filename
-    ).split('/')
-     .join('.')
-     .replace(/jsx?/, '..');
+    const prefixId = p
+      .relative(process.cwd(), file.opts.filename)
+      .split('/')
+      .join('.')
+      .replace(/jsx?/, '..');
 
     generatedDescriptors.push({ id: prefixId, defaultMessage });
   }
@@ -57,34 +66,74 @@ export default function ({types: t}) { // eslint-disable-line no-unused-vars
 
         exit(path, state) {
           const { file, opts, reactIntlMessages } = state;
-          const {basename, filename}    = file.opts;
+          const { basename, filename } = file.opts;
 
           let generatedDescriptors = reactIntlMessages.generatedDescriptors;
           file.metadata['react-intl-defineMessages'] = { generatedDescriptors };
 
           if (opts.messagesDir && generatedDescriptors.length > 0) {
-            let relativePath = p.join(p.sep, p.relative(process.cwd(), filename));
-            let generatedMessagesFilename = p.join(opts.messagesDir, p.dirname(relativePath), basename + 'Messages.js' );
+            let relativePath = p.join(
+              p.sep,
+              p.relative(process.cwd(), filename)
+            );
 
-            let namedDesriptors = generatedDescriptors.reduce((descriptorsWithKey, descriptor) => {
-              const lintFixedDescriptor = JSON.stringify(
-                descriptor, null, 4
-              )
-              .replace('}', '  }')
-              .replace('\"id\"', 'id')
-              .replace('\"defaultMessage\"', 'defaultMessage')
-              .replace(/\"/g, '\'');
+            let generatedMessagesFilename = p.join(
+              opts.messagesDir,
+              p.dirname(relativePath),
+              basename + 'Messages.js'
+            );
 
-              descriptorsWithKey += `  NameTheKey: ${lintFixedDescriptor},\n`;
-              return descriptorsWithKey;
-            }, '');
+            let existingContent;
+            if (existsSync(generatedMessagesFilename)) {
+              existingContent = readFileSync(
+                generatedMessagesFilename,
+                'utf-8'
+              );
+            }
 
-            let generatedDescriptorFile = `import { defineMessages } from 'react-intl';\n\nexport default defineMessages({\n${namedDesriptors}});`;
+            let namedDesriptors = generatedDescriptors.reduce(
+              (descriptorsWithKey, descriptor) => {
+
+                if (existingContent) {
+                  const { defaultMessage } = descriptor;
+                  const messageExist = new RegExp(
+                    `defaultMessage: '${defaultMessage}'`
+                  );
+
+                  if (messageExist.test(existingContent)) {
+                    return descriptorsWithKey;
+                  }
+                }
+
+                descriptorsWithKey += createDescriptorsWithKey(descriptor);
+                return descriptorsWithKey;
+              },
+              ''
+            );
+
             mkdirpSync(p.dirname(generatedMessagesFilename));
-            writeFileSync(generatedMessagesFilename, generatedDescriptorFile);
+
+            // updating existing file
+            if (existingContent && namedDesriptors) {
+              let oldDescriptors = removeDefindeMessageFormat(existingContent);
+
+              let updatedNamedDescriptors = oldDescriptors + namedDesriptors;
+              let updatedDescriptorFile = defineMessageFormat(
+                updatedNamedDescriptors
+              );
+
+              writeFileSync(generatedMessagesFilename, updatedDescriptorFile);
+            }
+
+            // new file
+            if (!existingContent && namedDesriptors) {
+              let generatedDescriptorFile = defineMessageFormat(namedDesriptors);
+              writeFileSync(generatedMessagesFilename, generatedDescriptorFile);
+            }
+            // else keep the file untouched
           }
-        },
-      },
+        }, // exit ends
+      }, // Program ends
 
       JSXText(path, state) {
         if (wasExtracted(path)) {
@@ -95,8 +144,8 @@ export default function ({types: t}) { // eslint-disable-line no-unused-vars
 
         // Tag the AST node so we don't try to extract it twice.
         tagAsExtracted(path);
-
       },
-    },
+
+    }, // visitor ends
   };
 }
